Fix vercel bug
